codex_artifacts/kits/mac-local-kit/requirements.txt
e2b==2.15.2
e2b-code-interpreter
psycopg[binary]==3.3.3
python-dotenv
requests==2.31.0
